fix(SB): move interfaces to fix demo warnings
