+++ packages/smart-board/package.json
@@ -85,12 +85,12 @@
   "limit-size": [
     {
       "path": "dist/index.min.js",
-      "limit": "256 Kb",
+      "limit": "384 Kb",
       "gzip": true
     },
     {
       "path": "dist/index.min.js",
-      "limit": "600 Kb"
+      "limit": "1024 Kb"
     }
   ]
 }
