+++ packages/smart-board/src/index.ts
@@ -1,4 +1,4 @@
 export * from './interfaces';
 export * from './components';
 export * from './SmartBoard';
-export { smartBoardConfig, ConfigObj, insights2Board } from './utils/smartBoardConfig';
+export { smartBoardConfig, insights2Board } from './utils/smartBoardConfig';
