+++ packages/smart-board/src/interfaces.ts
@@ -60,3 +60,57 @@ export type ChartCluster = Record<string, number>;
 
 /** The chart order */
 export type ChartOrder = Record<string, number>;
+
+export interface ConfigObj {
+  id?: string;
+  type: string;
+  data: any;
+  config: {
+    xField?: string;
+    yField?: string;
+    seriesField?: string;
+    isStack?: boolean;
+    isGroup?: boolean;
+    colorField?: string;
+    angleField?: string;
+  };
+  score?: number;
+  description?: string | string[];
+}
+
+export interface SmartBoardChartViewProps {
+  chartID: string;
+  chartInfo: any;
+  clusterID?: string;
+  interactionMode: string;
+  hasInsight: boolean;
+  hasLocked: boolean;
+  plotRender: (container: string | HTMLElement, type: any, data: any, options: any) => Object;
+  changeConnectionID: (string: string) => void;
+  quitResort: () => void;
+}
+
+export interface SmartBoardSamples {
+  sampleNames: string[];
+  initSampleIndex?: number;
+}
+
+export interface SmartBoardToolbarProps {
+  changeMode: (d: any) => void;
+  defaultMode?: string;
+}
+
+export interface SmartBoardSelectorProps {
+  changeSampleIndex: (d: any) => void;
+  samples: SmartBoardSamples;
+}
+
+export interface SmartBoardDashboardProps {
+  chartList: ChartListInfo;
+  chartGraph: ChartGraph;
+  chartOrder: ChartOrder;
+  chartCluster: ChartCluster;
+  interactionMode?: string;
+  hasInsight?: boolean;
+  plotRender: (container: string | HTMLElement, type: any, data: any, options: any) => Object;
+}
